move getDpsOutputs() helper into tt::ttir and a dialect-specific Utils.h
